+++ src/problem/func/func_LOAD_UTIL.c
@@ -18,7 +18,7 @@ Func* FUNC_LOAD_UTIL_new(REAL weight, Net* net) {
   FUNC_set_func_allocate(f, &FUNC_LOAD_UTIL_allocate);
   FUNC_set_func_clear(f, &FUNC_LOAD_UTIL_clear);
   FUNC_set_func_analyze_step(f, &FUNC_LOAD_UTIL_analyze_step);
-  FUNC_set_func_eval_setp(f, &FUNC_LOAD_UTIL_eval_step);
+  FUNC_set_func_eval_step(f, &FUNC_LOAD_UTIL_eval_step);
   FUNC_set_func_free(f, &FUNC_LOAD_UTIL_free);
   return f;
 }
@@ -39,7 +39,7 @@ void FUNC_LOAD_UTIL_clear(Func* f) {
   // Constant so not clear it
 
   // Counter
-  FUNC_set_Hcounter(f,0);
+  FUNC_set_Hphi_nnz(f,0);
 
   // Flags
   FUNC_clear_bus_counted(f);
@@ -52,7 +52,7 @@ void FUNC_LOAD_UTIL_count_step(Func* f, Branch* br, int t) {
   Bus* bus;
   Load* load;
   int bus_index_t[2];
-  int* Hcounter;
+  int* Hphi_nnz;
   char* bus_counted;
   int k;
   int T;
@@ -61,11 +61,11 @@ void FUNC_LOAD_UTIL_count_step(Func* f, Branch* br, int t) {
   T = BRANCH_get_num_periods(br);
 
   // Constr data
-  Hcounter = FUNC_get_Hcounter_ptr(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz_ptr(f);
   bus_counted = FUNC_get_bus_counted(f);
 
   // Check pointers
-  if (!Hcounter || !bus_counted)
+  if (!Hphi_nnz || !bus_counted)
     return;
 
   // Check outage
@@ -86,7 +86,7 @@ void FUNC_LOAD_UTIL_count_step(Func* f, Branch* br, int t) {
     if (!bus_counted[bus_index_t[k]]) {
       for (load = BUS_get_load(bus); load != NULL; load = LOAD_get_next(load)) {
 	if (LOAD_has_flags(load,FLAG_VARS,LOAD_VAR_P))
-	  (*Hcounter)++;
+	  (*Hphi_nnz)++;
       }
     }
 
@@ -99,10 +99,10 @@ void FUNC_LOAD_UTIL_allocate(Func* f) {
 
   // Local variables
   int num_vars;
-  int Hcounter;
+  int Hphi_nnz;
 
   num_vars = NET_get_num_vars(FUNC_get_network(f));
-  Hcounter = FUNC_get_Hcounter(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz(f);
 
   // gphi
   FUNC_set_gphi(f,VEC_new(num_vars));
@@ -110,7 +110,7 @@ void FUNC_LOAD_UTIL_allocate(Func* f) {
   // Hphi
   FUNC_set_Hphi(f,MAT_new(num_vars,
 			  num_vars,
-			  Hcounter));
+			  Hphi_nnz));
 }
 
 void FUNC_LOAD_UTIL_analyze_step(Func* f, Branch* br, int t) {
@@ -120,7 +120,7 @@ void FUNC_LOAD_UTIL_analyze_step(Func* f, Branch* br, int t) {
   Bus* bus;
   Load* load;
   int bus_index_t[2];
-  int* Hcounter;
+  int* Hphi_nnz;
   char* bus_counted;
   Mat* H;
   int k;
@@ -131,11 +131,11 @@ void FUNC_LOAD_UTIL_analyze_step(Func* f, Branch* br, int t) {
 
   // Constr data
   H = FUNC_get_Hphi(f);
-  Hcounter = FUNC_get_Hcounter_ptr(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz_ptr(f);
   bus_counted = FUNC_get_bus_counted(f);
 
   // Check pointers
-  if (!Hcounter || !bus_counted)
+  if (!Hphi_nnz || !bus_counted)
     return;
 
   // Check outage
@@ -156,10 +156,10 @@ void FUNC_LOAD_UTIL_analyze_step(Func* f, Branch* br, int t) {
     if (!bus_counted[bus_index_t[k]]) {
       for (load = BUS_get_load(bus); load != NULL; load = LOAD_get_next(load)) {
 	if (LOAD_has_flags(load,FLAG_VARS,LOAD_VAR_P)) {
-	  MAT_set_i(H,*Hcounter,LOAD_get_index_P(load,t));
-	  MAT_set_j(H,*Hcounter,LOAD_get_index_P(load,t));
-	  MAT_set_d(H,*Hcounter,2.*LOAD_get_util_coeff_Q2(load));
-	  (*Hcounter)++;
+	  MAT_set_i(H,*Hphi_nnz,LOAD_get_index_P(load,t));
+	  MAT_set_j(H,*Hphi_nnz,LOAD_get_index_P(load,t));
+	  MAT_set_d(H,*Hphi_nnz,2.*LOAD_get_util_coeff_Q2(load));
+	  (*Hphi_nnz)++;
 	}
       }
     }