+++ src/problem/func/func_REG_PHASE.c
@@ -18,7 +18,7 @@ Func* FUNC_REG_PHASE_new(REAL weight, Net* net) {
   FUNC_set_func_allocate(f, &FUNC_REG_PHASE_allocate);
   FUNC_set_func_clear(f, &FUNC_REG_PHASE_clear);
   FUNC_set_func_analyze_step(f, &FUNC_REG_PHASE_analyze_step);
-  FUNC_set_func_eval_setp(f, &FUNC_REG_PHASE_eval_step);
+  FUNC_set_func_eval_step(f, &FUNC_REG_PHASE_eval_step);
   FUNC_set_func_free(f, &FUNC_REG_PHASE_free);
   return f;
 }
@@ -39,19 +39,19 @@ void FUNC_REG_PHASE_clear(Func* f) {
   // Constant so not clear it
   
   // Counter
-  FUNC_set_Hcounter(f,0);
+  FUNC_set_Hphi_nnz(f,0);
 }
 
 void FUNC_REG_PHASE_count_step(Func* f, Branch* br, int t) {
 
   // Local variables
-  int* Hcounter;
+  int* Hphi_nnz;
 
   // Constr data
-  Hcounter = FUNC_get_Hcounter_ptr(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz_ptr(f);
 
   // Check pointer
-  if (!Hcounter)
+  if (!Hphi_nnz)
     return;
 
   // Check outage
@@ -59,17 +59,17 @@ void FUNC_REG_PHASE_count_step(Func* f, Branch* br, int t) {
     return;
   
   if (BRANCH_has_flags(br,FLAG_VARS,BRANCH_VAR_PHASE)) // phase var
-    (*Hcounter)++;
+    (*Hphi_nnz)++;
 }
 
 void FUNC_REG_PHASE_allocate(Func* f) {
   
   // Local variables
   int num_vars;
-  int Hcounter;
+  int Hphi_nnz;
   
   num_vars = NET_get_num_vars(FUNC_get_network(f));
-  Hcounter = FUNC_get_Hcounter(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz(f);
 
   // gphi
   FUNC_set_gphi(f,VEC_new(num_vars));
@@ -77,22 +77,22 @@ void FUNC_REG_PHASE_allocate(Func* f) {
   // Hphi
   FUNC_set_Hphi(f,MAT_new(num_vars,
 			  num_vars,
-			  Hcounter));
+			  Hphi_nnz));
 }
 
 void FUNC_REG_PHASE_analyze_step(Func* f, Branch* br, int t) {
 
   // Local variables
-  int* Hcounter;
+  int* Hphi_nnz;
   Mat* H;
   REAL dp;
 
   // Constr data
   H = FUNC_get_Hphi(f);
-  Hcounter = FUNC_get_Hcounter_ptr(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz_ptr(f);
 
   // Check pointer
-  if (!Hcounter)
+  if (!Hphi_nnz)
     return;
 
   // Check outage
@@ -105,10 +105,10 @@ void FUNC_REG_PHASE_analyze_step(Func* f, Branch* br, int t) {
     dp = FUNC_REG_PHASE_PARAM;
   
   if (BRANCH_has_flags(br,FLAG_VARS,BRANCH_VAR_PHASE)) { // phase var
-    MAT_set_i(H,*Hcounter,BRANCH_get_index_phase(br,t));
-    MAT_set_j(H,*Hcounter,BRANCH_get_index_phase(br,t));
-    MAT_set_d(H,*Hcounter,1./(dp*dp));
-    (*Hcounter)++;
+    MAT_set_i(H,*Hphi_nnz,BRANCH_get_index_phase(br,t));
+    MAT_set_j(H,*Hphi_nnz,BRANCH_get_index_phase(br,t));
+    MAT_set_d(H,*Hphi_nnz,1./(dp*dp));
+    (*Hphi_nnz)++;
   }
 }