+++ src/problem/func/func_REG_RATIO.c
@@ -18,7 +18,7 @@ Func* FUNC_REG_RATIO_new(REAL weight, Net* net) {
   FUNC_set_func_allocate(f, &FUNC_REG_RATIO_allocate);
   FUNC_set_func_clear(f, &FUNC_REG_RATIO_clear);
   FUNC_set_func_analyze_step(f, &FUNC_REG_RATIO_analyze_step);
-  FUNC_set_func_eval_setp(f, &FUNC_REG_RATIO_eval_step);
+  FUNC_set_func_eval_step(f, &FUNC_REG_RATIO_eval_step);
   FUNC_set_func_free(f, &FUNC_REG_RATIO_free);
   return f;
 }
@@ -39,19 +39,19 @@ void FUNC_REG_RATIO_clear(Func* f) {
   // Constant so not clear it
   
   // Counter
-  FUNC_set_Hcounter(f,0);
+  FUNC_set_Hphi_nnz(f,0);
 }
 
 void FUNC_REG_RATIO_count_step(Func* f, Branch* br, int t) {
 
   // Local variables
-  int* Hcounter;
+  int* Hphi_nnz;
 
   // Constr data
-  Hcounter = FUNC_get_Hcounter_ptr(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz_ptr(f);
 
   // Check pointers
-  if (!Hcounter)
+  if (!Hphi_nnz)
     return;
 
   // Check outage
@@ -59,11 +59,11 @@ void FUNC_REG_RATIO_count_step(Func* f, Branch* br, int t) {
     return;
   
   if (BRANCH_has_flags(br,FLAG_VARS,BRANCH_VAR_RATIO)) // ratio var
-    (*Hcounter)++;
+    (*Hphi_nnz)++;
 
   if (BRANCH_has_flags(br,FLAG_VARS,BRANCH_VAR_RATIO_DEV)) { // ratio dev var
-    (*Hcounter)++;
-    (*Hcounter)++;
+    (*Hphi_nnz)++;
+    (*Hphi_nnz)++;
   }
 }
 
@@ -71,10 +71,10 @@ void FUNC_REG_RATIO_allocate(Func* f) {
   
   // Local variables
   int num_vars;
-  int Hcounter;
+  int Hphi_nnz;
   
   num_vars = NET_get_num_vars(FUNC_get_network(f));
-  Hcounter = FUNC_get_Hcounter(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz(f);
 
   // gphi
   FUNC_set_gphi(f,VEC_new(num_vars));
@@ -82,22 +82,22 @@ void FUNC_REG_RATIO_allocate(Func* f) {
   // Hphi
   FUNC_set_Hphi(f,MAT_new(num_vars,
 			  num_vars,
-			  Hcounter));
+			  Hphi_nnz));
 }
 
 void FUNC_REG_RATIO_analyze_step(Func* f, Branch* br, int t) {
 
   // Local variables
-  int* Hcounter;
+  int* Hphi_nnz;
   Mat* H;
   REAL da;
 
   // Constr data
   H = FUNC_get_Hphi(f);
-  Hcounter = FUNC_get_Hcounter_ptr(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz_ptr(f);
 
   // Check pointer
-  if (!Hcounter)
+  if (!Hphi_nnz)
     return;
 
   // Check outage
@@ -110,23 +110,23 @@ void FUNC_REG_RATIO_analyze_step(Func* f, Branch* br, int t) {
     da = FUNC_REG_RATIO_PARAM;
   
   if (BRANCH_has_flags(br,FLAG_VARS,BRANCH_VAR_RATIO)) { // ratio var
-    MAT_set_i(H,*Hcounter,BRANCH_get_index_ratio(br,t));
-    MAT_set_j(H,*Hcounter,BRANCH_get_index_ratio(br,t));
-    MAT_set_d(H,*Hcounter,1./(da*da));
-    (*Hcounter)++;
+    MAT_set_i(H,*Hphi_nnz,BRANCH_get_index_ratio(br,t));
+    MAT_set_j(H,*Hphi_nnz,BRANCH_get_index_ratio(br,t));
+    MAT_set_d(H,*Hphi_nnz,1./(da*da));
+    (*Hphi_nnz)++;
   }
   
   if (BRANCH_has_flags(br,FLAG_VARS,BRANCH_VAR_RATIO_DEV)) { // ratio dev var
    
-    MAT_set_i(H,*Hcounter,BRANCH_get_index_ratio_y(br,t));
-    MAT_set_j(H,*Hcounter,BRANCH_get_index_ratio_y(br,t));
-    MAT_set_d(H,*Hcounter,1./(da*da));
-    (*Hcounter)++;
+    MAT_set_i(H,*Hphi_nnz,BRANCH_get_index_ratio_y(br,t));
+    MAT_set_j(H,*Hphi_nnz,BRANCH_get_index_ratio_y(br,t));
+    MAT_set_d(H,*Hphi_nnz,1./(da*da));
+    (*Hphi_nnz)++;
 
-    MAT_set_i(H,*Hcounter,BRANCH_get_index_ratio_z(br,t));
-    MAT_set_j(H,*Hcounter,BRANCH_get_index_ratio_z(br,t));
-    MAT_set_d(H,*Hcounter,1./(da*da));
-    (*Hcounter)++;
+    MAT_set_i(H,*Hphi_nnz,BRANCH_get_index_ratio_z(br,t));
+    MAT_set_j(H,*Hphi_nnz,BRANCH_get_index_ratio_z(br,t));
+    MAT_set_d(H,*Hphi_nnz,1./(da*da));
+    (*Hphi_nnz)++;
   }
 }