+++ src/problem/func/func_REG_SUSC.c
@@ -18,7 +18,7 @@ Func* FUNC_REG_SUSC_new(REAL weight, Net* net) {
   FUNC_set_func_allocate(f, &FUNC_REG_SUSC_allocate);
   FUNC_set_func_clear(f, &FUNC_REG_SUSC_clear);
   FUNC_set_func_analyze_step(f, &FUNC_REG_SUSC_analyze_step);
-  FUNC_set_func_eval_setp(f, &FUNC_REG_SUSC_eval_step);
+  FUNC_set_func_eval_step(f, &FUNC_REG_SUSC_eval_step);
   FUNC_set_func_free(f, &FUNC_REG_SUSC_free);
   return f;
 }
@@ -39,7 +39,7 @@ void FUNC_REG_SUSC_clear(Func* f) {
   // Constant so not clear it
 
   // Counter
-  FUNC_set_Hcounter(f,0);
+  FUNC_set_Hphi_nnz(f,0);
 
   // Flags
   FUNC_clear_bus_counted(f);
@@ -51,7 +51,7 @@ void FUNC_REG_SUSC_count_step(Func* f, Branch* br, int t) {
   Bus* bus[2];
   Shunt* shunt;
   int bus_index_t[2];
-  int* Hcounter;
+  int* Hphi_nnz;
   char* bus_counted;
   int k;
   int T;
@@ -60,11 +60,11 @@ void FUNC_REG_SUSC_count_step(Func* f, Branch* br, int t) {
   T = BRANCH_get_num_periods(br);
 
   // Constr data
-  Hcounter = FUNC_get_Hcounter_ptr(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz_ptr(f);
   bus_counted = FUNC_get_bus_counted(f);
 
   // Check pointers
-  if (!Hcounter || !bus_counted)
+  if (!Hphi_nnz || !bus_counted)
     return;
 
   // Check outage
@@ -86,11 +86,11 @@ void FUNC_REG_SUSC_count_step(Func* f, Branch* br, int t) {
       for (shunt = BUS_get_shunt(bus[k]); shunt != NULL; shunt = SHUNT_get_next(shunt)) {
 
 	if (SHUNT_has_flags(shunt,FLAG_VARS,SHUNT_VAR_SUSC)) // b var
-	  (*Hcounter)++;
+	  (*Hphi_nnz)++;
 
 	if (SHUNT_has_flags(shunt,FLAG_VARS,SHUNT_VAR_SUSC_DEV)) { // yz var
-	  (*Hcounter)++;
-	  (*Hcounter)++;
+	  (*Hphi_nnz)++;
+	  (*Hphi_nnz)++;
 	}
       }
     }
@@ -104,10 +104,10 @@ void FUNC_REG_SUSC_allocate(Func* f) {
 
   // Local variables
   int num_vars;
-  int Hcounter;
+  int Hphi_nnz;
 
   num_vars = NET_get_num_vars(FUNC_get_network(f));
-  Hcounter = FUNC_get_Hcounter(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz(f);
 
   // gphi
   FUNC_set_gphi(f,VEC_new(num_vars));
@@ -115,7 +115,7 @@ void FUNC_REG_SUSC_allocate(Func* f) {
   // Hphi
   FUNC_set_Hphi(f,MAT_new(num_vars,
 			  num_vars,
-			  Hcounter));
+			  Hphi_nnz));
 }
 
 void FUNC_REG_SUSC_analyze_step(Func* f, Branch* br, int t) {
@@ -124,7 +124,7 @@ void FUNC_REG_SUSC_analyze_step(Func* f, Branch* br, int t) {
   Bus* bus[2];
   Shunt* shunt;
   int bus_index_t[2];
-  int* Hcounter;
+  int* Hphi_nnz;
   char* bus_counted;
   Mat* H;
   int k;
@@ -136,11 +136,11 @@ void FUNC_REG_SUSC_analyze_step(Func* f, Branch* br, int t) {
 
   // Constr data
   H = FUNC_get_Hphi(f);
-  Hcounter = FUNC_get_Hcounter_ptr(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz_ptr(f);
   bus_counted = FUNC_get_bus_counted(f);
 
   // Check pointers
-  if (!Hcounter || !bus_counted)
+  if (!Hphi_nnz || !bus_counted)
     return;
 
   // Check outage
@@ -167,23 +167,23 @@ void FUNC_REG_SUSC_analyze_step(Func* f, Branch* br, int t) {
 
 	if (SHUNT_has_flags(shunt,FLAG_VARS,SHUNT_VAR_SUSC)) { // b var
 
-	  MAT_set_i(H,*Hcounter,SHUNT_get_index_b(shunt,t));
-	  MAT_set_j(H,*Hcounter,SHUNT_get_index_b(shunt,t));
-	  MAT_set_d(H,*Hcounter,1./(db*db));
-	  (*Hcounter)++;
+	  MAT_set_i(H,*Hphi_nnz,SHUNT_get_index_b(shunt,t));
+	  MAT_set_j(H,*Hphi_nnz,SHUNT_get_index_b(shunt,t));
+	  MAT_set_d(H,*Hphi_nnz,1./(db*db));
+	  (*Hphi_nnz)++;
 	}
 
 	if (SHUNT_has_flags(shunt,FLAG_VARS,SHUNT_VAR_SUSC_DEV)) { // yz var
 
-	  MAT_set_i(H,*Hcounter,SHUNT_get_index_y(shunt,t));
-	  MAT_set_j(H,*Hcounter,SHUNT_get_index_y(shunt,t));
-	  MAT_set_d(H,*Hcounter,1./(db*db));
-	  (*Hcounter)++;
+	  MAT_set_i(H,*Hphi_nnz,SHUNT_get_index_y(shunt,t));
+	  MAT_set_j(H,*Hphi_nnz,SHUNT_get_index_y(shunt,t));
+	  MAT_set_d(H,*Hphi_nnz,1./(db*db));
+	  (*Hphi_nnz)++;
 
-	  MAT_set_i(H,*Hcounter,SHUNT_get_index_z(shunt,t));
-	  MAT_set_j(H,*Hcounter,SHUNT_get_index_z(shunt,t));
-	  MAT_set_d(H,*Hcounter,1./(db*db));
-	  (*Hcounter)++;
+	  MAT_set_i(H,*Hphi_nnz,SHUNT_get_index_z(shunt,t));
+	  MAT_set_j(H,*Hphi_nnz,SHUNT_get_index_z(shunt,t));
+	  MAT_set_d(H,*Hphi_nnz,1./(db*db));
+	  (*Hphi_nnz)++;
 	}
       }
     }