+++ src/problem/func/func_SLIM_VMAG.c
@@ -18,7 +18,7 @@ Func* FUNC_SLIM_VMAG_new(REAL weight, Net* net) {
   FUNC_set_func_allocate(f, &FUNC_SLIM_VMAG_allocate);
   FUNC_set_func_clear(f, &FUNC_SLIM_VMAG_clear);
   FUNC_set_func_analyze_step(f, &FUNC_SLIM_VMAG_analyze_step);
-  FUNC_set_func_eval_setp(f, &FUNC_SLIM_VMAG_eval_step);
+  FUNC_set_func_eval_step(f, &FUNC_SLIM_VMAG_eval_step);
   FUNC_set_func_free(f, &FUNC_SLIM_VMAG_free);
   return f;
 }
@@ -39,7 +39,7 @@ void FUNC_SLIM_VMAG_clear(Func* f) {
   // Constant so not clear it
 
   // Counter
-  FUNC_set_Hcounter(f,0);
+  FUNC_set_Hphi_nnz(f,0);
 
   // Flags
   FUNC_clear_bus_counted(f);
@@ -51,7 +51,7 @@ void FUNC_SLIM_VMAG_count_step(Func* f, Branch* br, int t) {
   Bus* buses[2];
   Bus* bus;
   int bus_index_t[2];
-  int* Hcounter;
+  int* Hphi_nnz;
   char* bus_counted;
   int k;
   int T;
@@ -60,11 +60,11 @@ void FUNC_SLIM_VMAG_count_step(Func* f, Branch* br, int t) {
   T = BRANCH_get_num_periods(br);
 
   // Constr data
-  Hcounter = FUNC_get_Hcounter_ptr(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz_ptr(f);
   bus_counted = FUNC_get_bus_counted(f);
 
   // Check pointers
-  if (!Hcounter || !bus_counted)
+  if (!Hphi_nnz || !bus_counted)
     return;
 
   // Check outage
@@ -85,7 +85,7 @@ void FUNC_SLIM_VMAG_count_step(Func* f, Branch* br, int t) {
     if (!bus_counted[bus_index_t[k]]) {
 
       if (BUS_has_flags(bus,FLAG_VARS,BUS_VAR_VMAG)) // v var
-	(*Hcounter)++;
+	(*Hphi_nnz)++;
     }
 
     // Update counted flag
@@ -97,10 +97,10 @@ void FUNC_SLIM_VMAG_allocate(Func* f) {
 
   // Local variables
   int num_vars;
-  int Hcounter;
+  int Hphi_nnz;
 
   num_vars = NET_get_num_vars(FUNC_get_network(f));
-  Hcounter = FUNC_get_Hcounter(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz(f);
 
   // gphi
   FUNC_set_gphi(f,VEC_new(num_vars));
@@ -108,7 +108,7 @@ void FUNC_SLIM_VMAG_allocate(Func* f) {
   // Hphi
   FUNC_set_Hphi(f,MAT_new(num_vars,
 			  num_vars,
-			  Hcounter));
+			  Hphi_nnz));
 }
 
 void FUNC_SLIM_VMAG_analyze_step(Func* f, Branch* br, int t) {
@@ -117,7 +117,7 @@ void FUNC_SLIM_VMAG_analyze_step(Func* f, Branch* br, int t) {
   Bus* buses[2];
   Bus* bus;
   int bus_index_t[2];
-  int* Hcounter;
+  int* Hphi_nnz;
   char* bus_counted;
   Mat* H;
   int k;
@@ -129,11 +129,11 @@ void FUNC_SLIM_VMAG_analyze_step(Func* f, Branch* br, int t) {
 
   // Constr data
   H = FUNC_get_Hphi(f);
-  Hcounter = FUNC_get_Hcounter_ptr(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz_ptr(f);
   bus_counted = FUNC_get_bus_counted(f);
 
   // Check pointers
-  if (!Hcounter || !bus_counted)
+  if (!Hphi_nnz || !bus_counted)
     return;
 
   // Check outage
@@ -158,10 +158,10 @@ void FUNC_SLIM_VMAG_analyze_step(Func* f, Branch* br, int t) {
 	dv = FUNC_SLIM_VMAG_PARAM;
 
       if (BUS_has_flags(bus,FLAG_VARS,BUS_VAR_VMAG)) { // v var
-	MAT_set_i(H,*Hcounter,BUS_get_index_v_mag(bus,t));
-	MAT_set_j(H,*Hcounter,BUS_get_index_v_mag(bus,t));
-	MAT_set_d(H,*Hcounter,1./(dv*dv));
-	(*Hcounter)++;
+	MAT_set_i(H,*Hphi_nnz,BUS_get_index_v_mag(bus,t));
+	MAT_set_j(H,*Hphi_nnz,BUS_get_index_v_mag(bus,t));
+	MAT_set_d(H,*Hphi_nnz,1./(dv*dv));
+	(*Hphi_nnz)++;
       }
     }