+++ src/problem/func/func_SP_CONTROLS.c
@@ -18,7 +18,7 @@ Func* FUNC_SP_CONTROLS_new(REAL weight, Net* net) {
   FUNC_set_func_allocate(f, &FUNC_SP_CONTROLS_allocate);
   FUNC_set_func_clear(f, &FUNC_SP_CONTROLS_clear);
   FUNC_set_func_analyze_step(f, &FUNC_SP_CONTROLS_analyze_step);
-  FUNC_set_func_eval_setp(f, &FUNC_SP_CONTROLS_eval_step);
+  FUNC_set_func_eval_step(f, &FUNC_SP_CONTROLS_eval_step);
   FUNC_set_func_free(f, &FUNC_SP_CONTROLS_free);
   return f;
 }
@@ -39,7 +39,7 @@ void FUNC_SP_CONTROLS_clear(Func* f) {
   MAT_set_zero_d(FUNC_get_Hphi(f));
 
   // Counter
-  FUNC_set_Hcounter(f,0);
+  FUNC_set_Hphi_nnz(f,0);
 
   // Flags
   FUNC_clear_bus_counted(f);
@@ -53,7 +53,7 @@ void FUNC_SP_CONTROLS_count_step(Func* f, Branch* br, int t) {
   Gen* gen;
   Shunt* shunt;
   int bus_index_t[2];
-  int* Hcounter;
+  int* Hphi_nnz;
   char* bus_counted;
   int k;
   int T;
@@ -62,11 +62,11 @@ void FUNC_SP_CONTROLS_count_step(Func* f, Branch* br, int t) {
   T = BRANCH_get_num_periods(br);
 
   // Constr data
-  Hcounter = FUNC_get_Hcounter_ptr(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz_ptr(f);
   bus_counted = FUNC_get_bus_counted(f);
 
   // Check pointers
-  if (!Hcounter || !bus_counted)
+  if (!Hphi_nnz || !bus_counted)
     return;
 
   // Check outage
@@ -83,13 +83,13 @@ void FUNC_SP_CONTROLS_count_step(Func* f, Branch* br, int t) {
   if (BRANCH_is_tap_changer(br) &&
       BRANCH_has_flags(br,FLAG_VARS,BRANCH_VAR_RATIO) &&
       BRANCH_has_flags(br,FLAG_SPARSE,BRANCH_VAR_RATIO))
-    (*Hcounter)++;
+    (*Hphi_nnz)++;
 
   // Phase shift of phase-shifting transformer
   if (BRANCH_is_phase_shifter(br) &&
       BRANCH_has_flags(br,FLAG_VARS,BRANCH_VAR_PHASE) &&
       BRANCH_has_flags(br,FLAG_SPARSE,BRANCH_VAR_PHASE))
-    (*Hcounter)++;
+    (*Hphi_nnz)++;
 
   // Buses
   for (k = 0; k < 2; k++) {
@@ -102,7 +102,7 @@ void FUNC_SP_CONTROLS_count_step(Func* f, Branch* br, int t) {
       if (BUS_is_regulated_by_gen(bus) &&
 	  BUS_has_flags(bus,FLAG_VARS,BUS_VAR_VMAG) &&
 	  BUS_has_flags(bus,FLAG_SPARSE,BUS_VAR_VMAG))
-	(*Hcounter)++;
+	(*Hphi_nnz)++;
 
       // Generators
       for (gen = BUS_get_gen(bus); gen != NULL; gen = GEN_get_next(gen)) {
@@ -110,7 +110,7 @@ void FUNC_SP_CONTROLS_count_step(Func* f, Branch* br, int t) {
 	// Active power
 	if (GEN_has_flags(gen,FLAG_VARS,GEN_VAR_P) &&
 	    GEN_has_flags(gen,FLAG_SPARSE,GEN_VAR_P))
-	  (*Hcounter)++;
+	  (*Hphi_nnz)++;
       }
 
       // Shunts
@@ -120,7 +120,7 @@ void FUNC_SP_CONTROLS_count_step(Func* f, Branch* br, int t) {
 	if (SHUNT_is_switched(shunt) &&
 	    SHUNT_has_flags(shunt,FLAG_VARS,SHUNT_VAR_SUSC) &&
 	    SHUNT_has_flags(shunt,FLAG_SPARSE,SHUNT_VAR_SUSC))
-	  (*Hcounter)++;
+	  (*Hphi_nnz)++;
       }
     }
 
@@ -133,10 +133,10 @@ void FUNC_SP_CONTROLS_allocate(Func* f) {
 
   // Local variables
   int num_vars;
-  int Hcounter;
+  int Hphi_nnz;
 
   num_vars = NET_get_num_vars(FUNC_get_network(f));
-  Hcounter = FUNC_get_Hcounter(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz(f);
 
   // gphi
   FUNC_set_gphi(f,VEC_new(num_vars));
@@ -144,7 +144,7 @@ void FUNC_SP_CONTROLS_allocate(Func* f) {
   // Hphi
   FUNC_set_Hphi(f,MAT_new(num_vars,
 			  num_vars,
-			  Hcounter));
+			  Hphi_nnz));
 }
 
 void FUNC_SP_CONTROLS_analyze_step(Func* f, Branch* br, int t) {
@@ -155,7 +155,7 @@ void FUNC_SP_CONTROLS_analyze_step(Func* f, Branch* br, int t) {
   Gen* gen;
   Shunt* shunt;
   int bus_index_t[2];
-  int* Hcounter;
+  int* Hphi_nnz;
   char* bus_counted;
   Mat* H;
   int k;
@@ -166,11 +166,11 @@ void FUNC_SP_CONTROLS_analyze_step(Func* f, Branch* br, int t) {
 
   // Constr data
   H = FUNC_get_Hphi(f);
-  Hcounter = FUNC_get_Hcounter_ptr(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz_ptr(f);
   bus_counted = FUNC_get_bus_counted(f);
 
   // Check pointers
-  if (!Hcounter || !bus_counted)
+  if (!Hphi_nnz || !bus_counted)
     return;
 
   // Check outage
@@ -187,18 +187,18 @@ void FUNC_SP_CONTROLS_analyze_step(Func* f, Branch* br, int t) {
   if (BRANCH_is_tap_changer(br) &&
       BRANCH_has_flags(br,FLAG_VARS,BRANCH_VAR_RATIO) &&
       BRANCH_has_flags(br,FLAG_SPARSE,BRANCH_VAR_RATIO)) {
-    MAT_set_i(H,*Hcounter,BRANCH_get_index_ratio(br,t));
-    MAT_set_j(H,*Hcounter,BRANCH_get_index_ratio(br,t));
-    (*Hcounter)++;
+    MAT_set_i(H,*Hphi_nnz,BRANCH_get_index_ratio(br,t));
+    MAT_set_j(H,*Hphi_nnz,BRANCH_get_index_ratio(br,t));
+    (*Hphi_nnz)++;
   }
 
   // Phase shift of phase-shifting transformer
   if (BRANCH_is_phase_shifter(br) &&
       BRANCH_has_flags(br,FLAG_VARS,BRANCH_VAR_PHASE) &&
       BRANCH_has_flags(br,FLAG_SPARSE,BRANCH_VAR_PHASE)) {
-    MAT_set_i(H,*Hcounter,BRANCH_get_index_phase(br,t));
-    MAT_set_j(H,*Hcounter,BRANCH_get_index_phase(br,t));
-    (*Hcounter)++;
+    MAT_set_i(H,*Hphi_nnz,BRANCH_get_index_phase(br,t));
+    MAT_set_j(H,*Hphi_nnz,BRANCH_get_index_phase(br,t));
+    (*Hphi_nnz)++;
   }
 
   // Buses
@@ -212,9 +212,9 @@ void FUNC_SP_CONTROLS_analyze_step(Func* f, Branch* br, int t) {
       if (BUS_is_regulated_by_gen(bus) &&
 	  BUS_has_flags(bus,FLAG_VARS,BUS_VAR_VMAG) &&
 	  BUS_has_flags(bus,FLAG_SPARSE,BUS_VAR_VMAG)) {
-	MAT_set_i(H,*Hcounter,BUS_get_index_v_mag(bus,t));
-	MAT_set_j(H,*Hcounter,BUS_get_index_v_mag(bus,t));
-	(*Hcounter)++;
+	MAT_set_i(H,*Hphi_nnz,BUS_get_index_v_mag(bus,t));
+	MAT_set_j(H,*Hphi_nnz,BUS_get_index_v_mag(bus,t));
+	(*Hphi_nnz)++;
       }
 
       // Generators
@@ -223,9 +223,9 @@ void FUNC_SP_CONTROLS_analyze_step(Func* f, Branch* br, int t) {
 	// Active power
 	if (GEN_has_flags(gen,FLAG_VARS,GEN_VAR_P) &&
 	    GEN_has_flags(gen,FLAG_SPARSE,GEN_VAR_P)) {
-	  MAT_set_i(H,*Hcounter,GEN_get_index_P(gen,t));
-	  MAT_set_j(H,*Hcounter,GEN_get_index_P(gen,t));
-	  (*Hcounter)++;
+	  MAT_set_i(H,*Hphi_nnz,GEN_get_index_P(gen,t));
+	  MAT_set_j(H,*Hphi_nnz,GEN_get_index_P(gen,t));
+	  (*Hphi_nnz)++;
 	}
       }
 
@@ -236,9 +236,9 @@ void FUNC_SP_CONTROLS_analyze_step(Func* f, Branch* br, int t) {
 	if (SHUNT_is_switched(shunt) &&
 	    SHUNT_has_flags(shunt,FLAG_VARS,SHUNT_VAR_SUSC) &&
 	    SHUNT_has_flags(shunt,FLAG_SPARSE,SHUNT_VAR_SUSC)) {
-	  MAT_set_i(H,*Hcounter,SHUNT_get_index_b(shunt,t));
-	  MAT_set_j(H,*Hcounter,SHUNT_get_index_b(shunt,t));
-	  (*Hcounter)++;
+	  MAT_set_i(H,*Hphi_nnz,SHUNT_get_index_b(shunt,t));
+	  MAT_set_j(H,*Hphi_nnz,SHUNT_get_index_b(shunt,t));
+	  (*Hphi_nnz)++;
 	}
       }
     }
@@ -256,7 +256,7 @@ void FUNC_SP_CONTROLS_eval_step(Func* f, Branch* br, int t, Vec* var_values) {
   Gen* gen;
   Shunt* shunt;
   int bus_index_t[2];
-  int* Hcounter;
+  int* Hphi_nnz;
   char* bus_counted;
   REAL* phi;
   REAL* gphi;
@@ -276,11 +276,11 @@ void FUNC_SP_CONTROLS_eval_step(Func* f, Branch* br, int t, Vec* var_values) {
   phi = FUNC_get_phi_ptr(f);
   gphi = VEC_get_data(FUNC_get_gphi(f));
   Hd = MAT_get_data_array(FUNC_get_Hphi(f));
-  Hcounter = FUNC_get_Hcounter_ptr(f);
+  Hphi_nnz = FUNC_get_Hphi_nnz_ptr(f);
   bus_counted = FUNC_get_bus_counted(f);
 
   // Check pointers
-  if (!phi || !gphi || !Hd || !Hcounter || !bus_counted)
+  if (!phi || !gphi || !Hd || !Hphi_nnz || !bus_counted)
     return;
 
   // Check outage
@@ -313,8 +313,8 @@ void FUNC_SP_CONTROLS_eval_step(Func* f, Branch* br, int t, Vec* var_values) {
     gphi[index_val] = (1./sqrt_term)*((val-val0)/(dval*dval));
 
     // Hphi
-    Hd[*Hcounter] = FUNC_SP_CONTROLS_EPS/(dval*dval*sqrt_term*sqrt_term*sqrt_term);
-    (*Hcounter)++;
+    Hd[*Hphi_nnz] = FUNC_SP_CONTROLS_EPS/(dval*dval*sqrt_term*sqrt_term*sqrt_term);
+    (*Hphi_nnz)++;
   }
   else {
     // nothing because val0-val0 = 0 for constant val
@@ -340,8 +340,8 @@ void FUNC_SP_CONTROLS_eval_step(Func* f, Branch* br, int t, Vec* var_values) {
     gphi[index_val] = (1./sqrt_term)*((val-val0)/(dval*dval));
 
     // Hphi
-    Hd[*Hcounter] = FUNC_SP_CONTROLS_EPS/(dval*dval*sqrt_term*sqrt_term*sqrt_term);
-    (*Hcounter)++;
+    Hd[*Hphi_nnz] = FUNC_SP_CONTROLS_EPS/(dval*dval*sqrt_term*sqrt_term*sqrt_term);
+    (*Hphi_nnz)++;
   }
   else {
     // nothing because val0-val0 = 0 for constant val
@@ -374,8 +374,8 @@ void FUNC_SP_CONTROLS_eval_step(Func* f, Branch* br, int t, Vec* var_values) {
 	gphi[index_val] = (1./sqrt_term)*((val-val0)/(dval*dval));
 
 	// Hphi
-	Hd[*Hcounter] = FUNC_SP_CONTROLS_EPS/(dval*dval*sqrt_term*sqrt_term*sqrt_term);
-	(*Hcounter)++;
+	Hd[*Hphi_nnz] = FUNC_SP_CONTROLS_EPS/(dval*dval*sqrt_term*sqrt_term*sqrt_term);
+	(*Hphi_nnz)++;
       }
       else {
 	// nothing because val0-val0 = 0 for constant val
@@ -403,8 +403,8 @@ void FUNC_SP_CONTROLS_eval_step(Func* f, Branch* br, int t, Vec* var_values) {
 	  gphi[index_val] = (1./sqrt_term)*((val-val0)/(dval*dval));
 
 	  // Hphi
-	  Hd[*Hcounter] = FUNC_SP_CONTROLS_EPS/(dval*dval*sqrt_term*sqrt_term*sqrt_term);
-	  (*Hcounter)++;
+	  Hd[*Hphi_nnz] = FUNC_SP_CONTROLS_EPS/(dval*dval*sqrt_term*sqrt_term*sqrt_term);
+	  (*Hphi_nnz)++;
 	}
 	else {
 	  // nothing because val0-val0 = 0 for constant val
@@ -434,8 +434,8 @@ void FUNC_SP_CONTROLS_eval_step(Func* f, Branch* br, int t, Vec* var_values) {
 	  gphi[index_val] = (1./sqrt_term)*((val-val0)/(dval*dval));
 
 	  // Hphi
-	  Hd[*Hcounter] = FUNC_SP_CONTROLS_EPS/(dval*dval*sqrt_term*sqrt_term*sqrt_term);
-	  (*Hcounter)++;
+	  Hd[*Hphi_nnz] = FUNC_SP_CONTROLS_EPS/(dval*dval*sqrt_term*sqrt_term*sqrt_term);
+	  (*Hphi_nnz)++;
 	}
 	else {
 	  // nothing because val0-val0 = 0 for constant val